refactor(desktop): extract daily review main service
